FIXED: Slider's value now rounded correctly after changing the size (width) of the handle
[KTGUIEventProcessor]:
-added 'isDragged' variable;

[Slider]:
-added constrain [rangeStart...rangeEnd] for the 'value' variable being updated in 'updateValueFromHandlePosition()' method;
-changed the type of returned values from INTEGER to FLOAT for get/set methods of the 'handlePos', 'rangeStart', 'rangeEnd' and 'handleStep' variables (see above previous note);
-removed (int) casting from 'updateValueFromHandlePosition()', 'updateHandlePositionFromMouse()' and 'updateHandlePositionFromValue()' methods (see above two previous notes);
-ranamed 'handleWidth' variable to 'handleSize';
-renamed getHandleWidth() to getHandleSize();
-added 'isDragged' variable handling in 'processMouseDragged()' and 'processMouseReleased()';

[ScrollBar]:
-changed the type of returned values from INTEGER to FLOAT for getHandlePos(), getRangeStart() and setRangeEnd() methods;

+++ ktbyte-library/src/ktgui/ScrollBar.java
@@ -25,7 +25,7 @@ public class ScrollBar extends Controller {
 		slider.setValue(val);
 	}
 	
-	public int getHandlePos() {
+	public float getHandlePos() {
 		return slider.getHandlePos();
 	}
 	
@@ -45,7 +45,7 @@ public class ScrollBar extends Controller {
         slider.setIsValueVisible(visible);
     }
     
-	public int getRangeStart() {
+	public float getRangeStart() {
 		return slider.getRangeStart();
 	}
 	
@@ -53,7 +53,7 @@ public class ScrollBar extends Controller {
 		slider.setRangeStart(rangeStart);
 	}
 
-	public int getRangeEnd() {
+	public float getRangeEnd() {
 		return slider.getRangeEnd();
 	}
